// packages/database/convex/shared/generated_github_client.ts
export class ContentDirectory extends S.Array(
	S.Struct({
		type: S.Literal("dir", "file", "submodule", "symlink"),
		size: S.Int,
		name: S.String,
		path: S.String,
		content: S.optionalWith(S.String, { nullable: true }),
		sha: S.String,
		url: S.String,
		git_url: S.NullOr(S.String),
		html_url: S.NullOr(S.String),
		download_url: S.NullOr(S.String),
		_links: S.Struct({
			git: S.NullOr(S.String),
			html: S.NullOr(S.String),
			self: S.String,
		}),
	}),
) {}

export class ContentFileType extends S.Literal("file") {}

/**
 * Content File
 */
export class ContentFile extends S.Class<ContentFile>("ContentFile")({
	type: ContentFileType,
	encoding: S.String,
	size: S.Int,
	name: S.String,
	path: S.String,
	content: S.String,
	sha: S.String,
	url: S.String,
	git_url: S.NullOr(S.String),
	html_url: S.NullOr(S.String),
	download_url: S.NullOr(S.String),
	_links: S.Struct({
		git: S.NullOr(S.String),
		html: S.NullOr(S.String),
		self: S.String,
	}),
	target: S.optionalWith(S.String, { nullable: true }),
	submodule_git_url: S.optionalWith(S.String, { nullable: true }),
}) {}

export class ContentSymlinkType extends S.Literal("symlink") {}

/**
 * An object describing a symlink
 */
export class ContentSymlink extends S.Class<ContentSymlink>("ContentSymlink")({
	type: ContentSymlinkType,
	target: S.String,
	size: S.Int,
	name: S.String,
	path: S.String,
	sha: S.String,
	url: S.String,
	git_url: S.NullOr(S.String),
	html_url: S.NullOr(S.String),
	download_url: S.NullOr(S.String),
	_links: S.Struct({
		git: S.NullOr(S.String),
		html: S.NullOr(S.String),
		self: S.String,
	}),
}) {}

export class ContentSubmoduleType extends S.Literal("submodule") {}

/**
 * An object describing a submodule
 */
export class ContentSubmodule extends S.Class<ContentSubmodule>(
	"ContentSubmodule",
)({
	type: ContentSubmoduleType,
	submodule_git_url: S.String,
	size: S.Int,
	name: S.String,
	path: S.String,
	sha: S.String,
	url: S.String,
	git_url: S.NullOr(S.String),
	html_url: S.NullOr(S.String),
	download_url: S.NullOr(S.String),
	_links: S.Struct({
		git: S.NullOr(S.String),
		html: S.NullOr(S.String),
		self: S.String,
	}),
}) {}

export class ReposGetContent200 extends S.Union(
	ContentDirectory,
	ContentFile,
	ContentSymlink,
	ContentSubmodule,
) {}